Issue #3142986 by mangy.fox: Token entity is not checked before deletion

+++ src/OAuth2Storage.php
@@ -609,7 +609,11 @@ class OAuth2Storage implements OAuth2StorageInterface {
    */
   public function unsetRefreshToken($refresh_token) {
     $token = $this->getStorageToken($refresh_token);
-    $token->delete();
+
+    // Check token exists before trying to delete.
+    if ($token) {
+      $token->delete();
+    }
   }
 
   /**
